+++ Ryujinx.HLE/HOS/Ipc/IpcPtrBuffDesc.cs
@@ -4,11 +4,11 @@ namespace Ryujinx.HLE.HOS.Ipc
 {
     struct IpcPtrBuffDesc
     {
-        public long  Position { get; private set; }
-        public int   Index    { get; private set; }
-        public long  Size     { get; private set; }
+        public ulong Position { get; private set; }
+        public uint  Index    { get; private set; }
+        public ulong Size     { get; private set; }
 
-        public IpcPtrBuffDesc(long position, int index, long size)
+        public IpcPtrBuffDesc(ulong position, uint index, ulong size)
         {
             Position = position;
             Index = index;
@@ -17,20 +17,20 @@ namespace Ryujinx.HLE.HOS.Ipc
 
         public IpcPtrBuffDesc(BinaryReader reader)
         {
-            long word0 = reader.ReadUInt32();
-            long word1 = reader.ReadUInt32();
+            ulong word0 = reader.ReadUInt32();
+            ulong word1 = reader.ReadUInt32();
 
-            Position  =  word1;
+            Position =  word1;
             Position |= (word0 << 20) & 0x0f00000000;
             Position |= (word0 << 30) & 0x7000000000;
 
-            Index  = ((int)word0 >> 0) & 0x03f;
-            Index |= ((int)word0 >> 3) & 0x1c0;
+            Index  = ((uint)word0 >> 0) & 0x03f;
+            Index |= ((uint)word0 >> 3) & 0x1c0;
 
             Size = (ushort)(word0 >> 16);
         }
 
-        public IpcPtrBuffDesc WithSize(long size)
+        public IpcPtrBuffDesc WithSize(ulong size)
         {
             return new IpcPtrBuffDesc(Position, Index, size);
         }
@@ -42,8 +42,8 @@ namespace Ryujinx.HLE.HOS.Ipc
             word0  = (uint)((Position & 0x0f00000000) >> 20);
             word0 |= (uint)((Position & 0x7000000000) >> 30);
 
-            word0 |= (uint)(Index & 0x03f) << 0;
-            word0 |= (uint)(Index & 0x1c0) << 3;
+            word0 |= (Index & 0x03f) << 0;
+            word0 |= (Index & 0x1c0) << 3;
 
             word0 |= (uint)Size << 16;
